remove Main.css and rename app.js
